+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * [server] Introduce `Client::credentials()` to allow retrieving the credentials of client socket (implements `wl_client_get_credentials`).
+* [protocols] Introduce misc/input-method-unstable-v2
 
 ## 0.29.0 -- 2021-07-24