+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
+## 0.29.1 -- 2021-10-17
+
 #### Additions
 
 - [server] Introduce `Client::credentials()` to allow retrieving the credentials of client socket (implements `wl_client_get_credentials`).
 - [protocols] KDE's `server-decoration` protocol was added to misc module.
-* [protocols] Introduce misc/input-method-unstable-v2
+- [protocols] Introduce misc/input-method-unstable-v2
 
 ## 0.29.0 -- 2021-07-24